[#183] typing - 타입 파일 분리
