+++ backend/src/types/auth.ts
@@ -1,0 +1,9 @@
+export interface AccessInfo {
+    username: string
+    roles: string[]
+}
+
+export interface StrictAccessInfo {
+    username: string
+    is_edit_token: boolean
+}
