+++ backend/src/types/file.ts
@@ -1,0 +1,12 @@
+export interface FileInfo {
+    id: string
+    filename: string
+    size: number
+    uploader: string
+    timestamp: Date
+}
+
+export interface FileXRef {
+    target: string
+    ref: string
+}
