+++ backend/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './api'
+export * from './auth'
+export * from './file'
